Refactored to visually simplify names based on comments.
