+++ README.md
@@ -1,5 +1,5 @@
 # BANZAI-FLOYDS-UI
-Frontend for re-extraction using BANZAI-FLOYDS
+User interface for re-extraction using BANZAI-FLOYDS
 
 Installation
 ------------
@@ -23,3 +23,8 @@ docker run --rm -p 8080:8080 banzai-floyds-ui
 
 And then point your browser to [http://127.0.0.1:8080/](http://127.0.0.1:8080/) for the API root of the project.
 The main dashboard can be found at [http://127.0.0.1:8080/banzai-floyds](http://127.0.0.1:8080/banzai-floyds).
+
+Development
+-----------
+This package is broken into two subpackages, one for the API, and one for the GUI (web frontend). The API is a standard Django
+Rest Framework. The GUI is comprised of a Django Plotly Dash app to enable user interaction.
