+++ banzai_floyds_ui/api/urls.py
@@ -1,5 +1,5 @@
 from django.urls import path
-from banzai_floyds_ui.banzai_floyds_api import views
+from banzai_floyds_ui.api import views
 
 urlpatterns = [
     path('', views.printHelp)
