+++ banzai_floyds_ui/gui/scaffold.py
@@ -1,4 +1,4 @@
-import banzai_floyds_ui.banzai_floyds_gui.banzai_floyds_app as banzai_floyds_app
+import banzai_floyds_ui.gui.app
 import logging
 logger = logging.getLogger(__name__)
 
@@ -6,6 +6,6 @@ logger = logging.getLogger(__name__)
 def loader(app_name):
     if 'banzai-floyds' in app_name:
         logger.info(f"{app_name} is being loaded")
-        return banzai_floyds_app.app
+        return banzai_floyds_ui.gui.app.app
     else:
         return None
