+++ banzai_floyds_ui/gui/urls.py
@@ -15,12 +15,12 @@ Including another URLconf
 """
 from django.contrib import admin
 from django.urls import path, include
-from banzai_floyds_ui.banzai_floyds_gui.views import banzai_floyds_view, login_view
+from banzai_floyds_ui.gui.views import banzai_floyds_view, login_view
 
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('', include('banzai_floyds_ui.banzai_floyds_api.urls')),
+    path('', include('banzai_floyds_ui.api.urls')),
     path('banzai-floyds', banzai_floyds_view, name="banzai-floyds"),
     path('login', login_view, name="login"),
     path('django_plotly_dash', include('django_plotly_dash.urls')),
