+++ pyproject.toml
@@ -7,7 +7,7 @@ include = ["banzai_floyds_ui*"]
 
 [tool.setuptools.package-data]
 "banzai_floyds_ui" = ["static/*"]
-"banzai_floyds_ui.banzai_floyds_gui" = ["templates/*"]
+"banzai_floyds_ui.gui" = ["templates/*"]
 
 [project]
 dynamic = ["version"]
